refactor(#792): move runtime_audio_frame inline tests to a sibling file (sweep)
runtime_audio_frame.rs was 743 lines — ~380 of them an inline #[cfg(test)]
mod elastic_tests (the 34 ElasticBuffer/AudioFrame invariant tests). Moved the
module to runtime_audio_frame_tests.rs, wired via #[cfg(test)] #[path] mod
(the repo convention); production drops to 364 and leaves the debt list. Pure
move — all 34 tests pass from the new location, zero warnings.
